Stop pinning mozbase dependencies
Normally pinning dependencies is a very good idea. But in this case, we
will be moving raptor into mozilla-central (where it will use the latest
mozbase versions). Therefore, we want to make sure this prototype works
with the latest versions of mozbase as possible to reduce the risk of
things going wrong when we merge it to mozilla-central.

While unpinning only gives us the latest release (not the
mozilla-central tip), it's still better than being pinned to an old
version.