Fix the 'get_binary' fixture
The test_start_browser function still isn't running, but at least the
'get_binary' fixture now works.

It will try to find a binary by:
1) Looking for a local build (only relevant within mozilla-central)
2) Looking for a previously setup binary in the temporary directory
3) Looking for a GECKO_INSTALLER_URL env variable
4) Looking for a GECKO_BINARY_PATH env variable

When testing locally, option 4) is the best bet, as it is quick and easy
to control. If no binary could be detected with any of those methods the
test will be skipped.

There is currently no support for Chrome (thus why it xfails instead of
skips).