Always use up-to-date libraries in the vendor directory
This makes it so running 'pip install -r mozbase_requirements.txt' is no
longer necessary. This was a problem because it meant changes to the
vendor directory weren't picked up unless you re-installed the packages
in your virtualenv. Due to pip's caching, this was a pain to do.

Now we simply add the 'vendor' directory to sys.path directly in the
'raptor' module, so we'll pick up changes automatically. W'll need to
remember to remove this when we land in mozilla-central.